requirement.txt
absl-py==0.10.0
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.2.0
behave==1.2.6
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
deepclustering2 @ file:///home/jizong/Workspace/release-code/deepclustering2-2.0.0-py3-none-any.whl
easydict==1.9
filelock==3.0.12
future==0.18.2
gdown==3.12.2
google-auth==1.22.0
google-auth-oauthlib==0.4.1
GPUQueue==0.0.3
grpcio==1.32.0
idna==2.10
imageio==2.9.0
importlib-metadata==2.0.0
joblib==0.16.0
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.3.2
MedPy==0.4.0
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
msgpack==1.0.0
multidict==4.7.6
networkx==2.5
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1596233707986/work
oauthlib==3.1.0
olefile==0.46
opencv-python==4.4.0.44
pandas==1.1.2
parse==1.18.0
parse-type==0.5.2
Pillow @ file:///tmp/build/80754af9/pillow_1594307325547/work
protobuf==3.13.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytorch-ranger==0.1.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.2
SimpleITK==2.0.0
six==1.15.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.1
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2020.10.1
torch==1.6.0
torch-optimizer==0.0.1a15
torchvision==0.7.0
tqdm==4.50.0
typing-extensions==3.7.4.3
urllib3==1.25.10
Werkzeug==1.0.1
yarl==1.6.0
zipp==3.2.0
kornia
